Reorder macros, so no revision for unchanged pages
Confluence detects that pages are identical, and does not generate
a new revision, but this was broken because macros were stored in
Confluence markup differently to how they were being uploaded.

Ordering macros in the same way that Confluence rewrites them, and putting
no newlines in between arguments, means that uploading an unchanged page
does not trigger a new revision.