Require call syntax to resolve resource URLs
Some data values are in the form of URLs, which can be retrieved to obtain
more information.  Previously, these looked just like normal attribute
access, as though the entire record had been downloaded at once. However,
it was also useful to have the URL string sometimes.  So these URLs were
contained in a special class that mimicked both a string and a resource.

To simplify things, this patch converts them to a subclass of string, so
they are mostly a string, but using the call syntax 's()' provides the
remote resource contents.  This makes it more explicit when a remote call
is being done, and actually clarifies what is happening, since at these
points there is often repetition, when an attribute access provides the URL
and the top-level of the retrieved results also contains the same attribute
name.